refactor(mantine): extract table styles into its own file
